+++ tests/testthat/test-unnest-tokens.R
@@ -177,11 +177,13 @@ test_that("unnest_tokens raises an error if custom tokenizer gives bad output", 
 
   expect_error(
     unnest_tokens(d, word, txt, token = function(e) c("a", "b")),
-    "to be a list"
+    "to be a list",
+    class = "rlang_error"
   )
   expect_error(
     unnest_tokens(d, word, txt, token = function(e) list("a", "b")),
-    "of length"
+    "of length",
+    class = "rlang_error"
   )
 })