+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Update tidying test for new tibble release (inner names for columns)
 - Deprecate SE versions of main functions (have long been replaced by tidy eval semantics)
+- Improve error handling throughout
 
 # tidytext 0.2.3